graphfc/requirements.txt
openai>=1.0.0
transformers>=4.25.0
torch>=1.13.0
datasets>=2.14.0
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.3.0
tqdm>=4.64.0
jsonlines>=3.1.0
requests>=2.28.0
matplotlib>=3.6.0
seaborn>=0.12.0
pyserini>=0.20.0
rank_bm25>=0.2.2
nltk>=3.8.0
spacy>=3.4.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
pyyaml>=6.0
python-dotenv>=1.0.0
wandb>=0.15.0
accelerate>=0.20.0
sentence-transformers>=2.2.0
google-generativeai>=0.8.0